更新:model/User.age,entity/UserTable.age 型longからintへ変更
影響クラス:HomeController、FinishController、User関連、各テストクラス

+++ src/main/java/com/example/demo/mapper/UserMapper.java
@@ -16,7 +16,7 @@ public interface UserMapper {
 	 * find max user no
 	 * @return
 	 */
-	public int findMaxNo();
+	public long findMaxNo();
 	
 	/**
 	 * find one user
